Set Up File Structure
I started using some pretty cool features in Jekyll and figured out my
SASS workflow. So, now I've got all my pages in the root.

The pages will be partially populated with includes (I'm going to write
one big one for all the SEO, fonts, etc.). All the styling will be done
from the stylesheet in the root, which will be compiled from the
partials in the styles folder.

Before starting, I set up a detached Jekyll server and then set it to
watch for changes in the background. I also set up a Sass --watch
operation that watches for changes in base.scss (which imports all the
partials). That way, I can preview the site before publishing it.

I need to remember to set up Netlify to run a jekyll build command and
host from the _Site directory.