Added Hamburger/Nav Menu and Button
Did my first ever CSS3 animations to create a hamburger menu, and laid out the navigation buttons.
The navigation menu has a short script that uses Jekyll to color the active menu item red.
Doing the animations for the hamburger icon were more difficult than expected, I had to force trigger a reflow on the function attached to the click event, in order to get the animations to restart. It all worked out though, and looks slick now.