Linked up Nav and Added Keypress Event
The page is now primed for content! I added a keypress event for the
menu (pushing me opens it up now). I also linked up the navigation bar
to all the pages, and updated all the YAML Front Matter.

The site is functioning beautifully, and ready for a rapid iteration
cycle.