Finalized Homepage, Lots of Tweaks
The biggest changes include a "play reel" button, working footer, and social media icons.

I also tweaked the navMenuButton so that it doesn't disappear when you scroll through a page.

There have also been many other tweaks. The "X" doesn't animate in on every page load anymore, the navbuttons are underscored instead of highlighted in white, the images loaded in the background are different, the privacy policy and terms & conditions buttons work, the 404 error page is styled, etc.