Significant Under-The-Hood Changes to News Section
Re-wrote myNews.js to be WAY BETTER. Now, the page uses unique '#' urls
for each post. Click events are no longer as important, I'm now using
event listeners for hash changes.

This does a few things. First, it allows you to refresh the page without
resetting to the default news page. Second, it allows news stories to be
share-able. This was a little trickier on mobile, because although it
would open the news story, the story could be way down on the page. SO I
added some automatic scrolling.

Eventually I'll have to figure out how to limit the number of posts on a
page...but that's a problem for another day.