Finished Contact Page, Started About Page
The biggest change is the new contact page, which has a stylized Google
Map embed (with custom marker) and an info card with websites/address
etc. I also started work on the embed page, and tweaked things across
the site (including changing the hue of the red slightly to match the
theme better).

I also added a start.command batch file to the directory (and added it
to the .gitignore) so that I can get up and running to build this thing
faster.